pyproject.toml
[project]
name = "django-vs-litestar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "click>=8.3.2",
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.27",
    "httpx>=0.28.1",
    "litestar>=2.21.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "python-json-logger>=3.3",
    "sgp4>=2.25",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.12.0",
    "respx>=0.23.1",
    "testcontainers[postgres]>=4.14.2",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_api.config.settings"
asyncio_mode = "auto"
pythonpath = ["."]
